Register types and type methods as symbols
They are in the form of _type.${type} and _type.${type}.${method}, with
a clear ancestry below the top-level _type symbol. The SYM_METHOD kind of
symbol was already present but never used so it just found its purpose.

They appear at the beginning of the symbol table:

    /*
     * Symbol Table
     *
     * reserved  VOID       40 41 _type
     * none      VOID       40 41 _type.STEVEDORE
     * method    BYTES      40 41 _type.STEVEDORE.free_space
     * method    BOOL       40 41 _type.STEVEDORE.happy
     * method    BYTES      40 41 _type.STEVEDORE.used_space
     * none      VOID       40 41 _type.STRINGS
     * method    STRING     40 41 _type.STRINGS.lower
     * method    STRING     40 41 _type.STRINGS.upper
     * reserved  VOID       41 41 acl
     * reserved  VOID       41 41 backend
     * [...]
     */

Unlike VMOD functions or object methods, native type methods are never
invoked as a standalone statement:

    strings.upper();

They are only evaluated atop an expression:

    (string expression).upper();

So any VMOD named after a type, like vmod_blob, should not conflict with
native type methods in the symbol table.