Fix WS_ReserveSize calls when bytes is equal to free workspace
Currently, with the 505b7bd9643006fa8e3977f920564ce12a2b24a2 patch, when
calling WS_ReserveSize with bytes equal to the amount of workspace that is
currently available, it will return zero and mark overflow.

This patch redoes the patch, and changes it to return zero and overflow
only when the requested number of bytes is larger than what is available.