Use absolute paths in dlopen(3) for VCLs and VMODs
This improves the out-of-the-box debugging experience for at least gdb
and lldb, allowing them to grab type information for the symbols coming
from such shared objects.

This however doesn't help further when multiple VCLs or VMODs define the
same symbols, for example in the case of similar VCLs (reload scenario)
or a VMOD upgrade (either with import+from or VMOD caching kicking in).

For the latter case, #2966 would help.