demo and test use of the stringified session attributes
While, in most of the code, we now use SA_*_ADDR and format the
address/port when needed, for the client ip and port, there still exist
formatted strings as session attributes.

I recently came across this in a different context and wondered if we
should remove these string session attributes, but as we use them multiple
times for each request, that would only imply additional overhead for
repeated string formatting.  So I think we should keep them for now,
but we might want to consider to add to struct suckaddr optional
formatted strings to avoid the duplicated formatting which already
happens.

This commit adds a regression test.

Motivated by #3173