Restore previous VRT_AddDirector semantics
Upon failure, don't destroy the underlying director implementation. When
the deletion code was split in c671bb6691f27172ae9103226cb09a99c9812483
in order to be reused in 615298547adfb06aa785b42552062bc391f56ecd, it
mistakenly included the director implementation.

Instead, it's really the caller's job to tear down the backend if it
couldn't be added. Since vbe_destroy operates on a director, it is split
in two to avoid reintroducing the previous "undo" code duplication.

Refs #3176