Allow acceptor tasks to bail out earlier
This is an attempt at solving the fundamental problem with thread pool
removal. In order to avoid a situation where an acceptor task might be
"stuck" forever in an accept call we poll the socket to get a periodic
chance to check the pool status.

To avoid the race between poll(2) and accept(2) we need the socket to
be non-blocking at this moment. If two acceptors compete for the same
connection one of them may go back to polling the socket.

This fixes the two memory leaks reported by ASAN for c80.

Refs 656982a5cf7042a8576c2e9d48defcd25d749fbd