Plug leaks caught by c80 with ASAN
This is only a matter of leaving the acceptor loop to always free the
poolsock before returning from the function. This was initially done in
only one place out of three, with no room to exit the accept inner loop
in the absence of an incoming connection.

With this we may have everything in place to drop the drop_pools debug
flag.

Refs 656982a5cf70
Closes #3177