Polish VTCP usage in the acceptor setup
This adds the errno value to error logs for TCP fast open and accept
filters. For FreeBSD errno is cleared prior to calling setsockopt with
the hope that it will help understand if it's actually failing with an
undocumented ENOENT.

Speaking of documentation, the return value is either 0 or -1 for so
there's no point in logging that.