Tweak test cases to work with or disable accept_filter
The test cases disabling the accept_filter parameter are those involving
GET requests with a body that are meant to be cached, and one test case
covering OOB data. The rest uses the POST method since the httpready
filter will only let syntactically correct GET or HEAD methods pass to
the application.

Apparently FreeBSD's httpready filter considers that a GET request with
a body is syntactically incorrect although this is not specifically said
in the manual. The HTTP specification doesn't forbid such requests and
they are hard to justify considering the semantics of GET and HEAD, but
there are products in the wild relying on that.

On the other hand GET\r\n\r\n isn't considered malformed, see r01881.vtc.

We don't need the dataready filter on FreeBSD for listen addresses that
expect a PROXY protocol header, the httpready filter will effectively
work like that if the request doesn't look like a GET or a HEAD.