Attempt at stabilizing s10
The stabilization attempt still fails under heavy load on FreeBSD
and the idle_send_timeout may not kick in. However we now make sure
that the client and varnish have a weak synchronization point for
headers and that only the large body (compared to the socket buffers)
is subject to the timeouts. We also avoid running into EPIPE in
varnish when we miss idle_send_timeout but that will in turn trigger
a very slow deliver.

This test has been relying on SLT_Debug records from day one and now
that we have SLT_Notice we could perpetuate this information and at
the same time grant ourselves the freedom to explain each case and
which parameters may be used to try to improve the situation.