Improve workspace_reserve test
The test is now conducted on the session workspace in order to have the
client workspace free to generate vtc.log() output which needs
workspace to format the string.

This made me notice that vtc.log, via WS_Reset(), clears a previous
workspace overflow and I wonder if this is how it should be.