straighten cli ctx vs. temperature event
Our code was inconsistent with respect to the cli vrt_ctx.msg being set
for vcl temperature events.

This patch is intended as a temporary bandaid to at least make the code
consistent.

A proper solution is pending after clarification of the intent in #2902,
in particular wrt VCL_EVENT_COLD, which, by design, should not have a
ctx->msg, but does have one at the moment.

See #2902 for more details