Add a VRE_quote() function
This is inspired by PHP's preg_quote() that is used to escape characters
that may have a special meaning for PCRE depending on their location in
the subject string.

This allows to build regular expressions from a mix of trusted strings
and arbitrary user input. There is unfortunately no such function in
PCRE.