Introduce a struct gethdr_t type
This replaces the (struct gethdr_s).what string, making it explicit in
the type system that the first octet of such strings contains a length
without changing the existing memory layout.

A convenience macro is here to create such a struct from a constant
string literal, which allows in turn to create true header constants
and drop the HTTP_Init() step.

The diff may be easier to follow with the --word-diff option.