Add VRT_AllocStrandsWS() to allocate strands on a workspace
So far, we have focused on transitioning vmod arguments to STRANDS.

To take the next step and also return STRANDS (which leverages the
actual benefit - not copying unmodified strand elements), vmods need
to allocate a struct strands on the workspace.

This commit adds a utility function for this common task.

Implementation note: A single WS_Alloc() with some pointer arithmetic
would suffice, but using two results in cleaner code.