Enable automake maintainer mode for development
In maintainer mode, whenever a file managed by autoconf or automake is
changed it is regenerated on the fly. For Makefile.am files this also
applies to any fragment .am file referenced by the include directive.

The configure script disables the maintainer mode by default, which is
probably a sane behavior because downstream usage of dist archives don't
need the various autotools to be present, unless a Makefile or similar
is patched.

This is explicitly enabled by the autogen.des script, in order to make
this change only impact development, saving most of the time a configure
step after amending something in the build.

Refs #3184