Ensure everything is built before running tests
With a growing list of bundled VMODs we want to split the main test
suite to colocate VMOD-specific tests with the rest of its sources.
Because a VMOD test case may involve other binaries, and will always
involve at least varnishtest and varnishd, we insist that make check
kicks in only after a complete recursive build.

This doesn't work in this scenario:

    cd lib/libvmod_<name> && make check

The assumption is that at this point a full build was already performed
and changes are only happening in that sub directory.

Refs #3184