Generate test suite setup in vmodtool
That part of the automake boilerplate is generated conditionally, and
since we check automake_boilerplate.am files in the list of tests is
printed with one per line to be diff-friendly.

A test was "randomly" chosen to be the first VMOD-specific test.

Generated automake_boilerplate.am files no longer start nor end with
empty lines.

This is probably going to mandate some changes in vtest to collect
VMOD-specific test suites.

Refs #3184