Initial release notes for Varnish 6.3.0
This is a result of a commit review between 6.2.0 and 6.3.0 and while
it should in theory be comprehensive, it is not. The upgrade notes in
particular are developer-centric, which may be a sign that none of the
changes require attention during an upgrade it could as well be an
oversight.

There is one item that I identified that should be mentioned in the
upgrade guide. I invite @nigoroll to give a shot at the VCL auto state.

In the absence of a fair warning, I extended WS_Reserve's grace period.

I omitted the -rcvbuf action on purpose, for the same reason why it is
not documented in vtc(7) on purpose: it's currently used in conjunction
with vmod_debug and we don't ship it.

While this is a milestone, I don't consider this work to be done so I
didn't remove the notes saying so.

We should also probably partially pick 53b4d2f58202 on the 6.3 branch.