Collapse body_status and req_body_status into one type.
A lot of this is s/REQ_BODY_/BS_/g, followed by a reduction
pass to go from testing on specific status values to attributes
of the status value.

For instance the req.body caching code does not need to know if
how the transport sees this, chunked vs. eof for instance, it just
needs to know that there is a body and that we do not know the
length of it yet.

The transports own the "vocabulary", because internally H1 needs
to know the difference between BS_CHUNKED and BS_EOF in order to
setup the VFPs.