Keep Age information on passes
When doing a pass, we would remove the Age header from the backend, and
create a new one based on the time the fetch was initiated. This creates
problems when calculating the time to live in downstream caches (browser
cache or layered varnishes).

With this patch, the RFC_2616_Ttl calculation routine is run also for
passes, where the t_origin field of the object is adjusted for an incoming
Age header. This makes sure that the Age header generated during delivery
is correct. The rest of the Ttl calculation is skipped for passes,
including the logging of SLT_TTL "RFC".

Fixes: varnishcache/varnish-cache#3221