RFC: Expansion of probe .request definitions
It is possible to share a probe between multiple backends today, but it
becomes impossible to share a probe with a .request field with a proper
Host header.

Either the backend is in cahoots with Varnish and will expect a specific
domain for health checks (for example Host: health) or we simply don't
send the header despite its mandatory status in HTTP/1.1.

This patch introduces incomplete and rudimentary %{variable} expansion
in the probe request in a way that could be outsourced in an include
table used both to build a probe request or generate documentation.

An initial set of two variables is implemented:

- %{host_header} for the reason mentioned above
- %{now} in order to generate a Date header

We could imagine more variables like %{version} as suggested by the test
case. The %{...} syntax was selected to avoid conflicts with varnishtest
macros and is only here to illustrate the mechanics. This is something I
would tend to check at vcl.load time, this is completely missing from
this RFC.