Kill a dead branch when building a probe request
We have two fallbacks in the absence of a .host_header field in the
backend, one of which is systematic. We can build a non .request probe
request with a single printf statement.

Please note that when the .host field is in the form "$addr $port" we
use that as the probe's host header. This behavior is not introduced by
this change but ultimately we might want to do something about it and
for example capture the first token from hosthdr with a txt.