Move def and units fields up in struct parspec
It's a bit weird that the default value and the unit is separated from
the min and max. This aligns struct parspec field order with the PARAM()
macro for these two fields.

In static parspec definitions, the flags field can now be omitted if
there are no flags, and the parameter is not subject to dynamic bounds.