Speed up Travis CI builds
The longest part is the test suite, so in order to release build
machines sooner I hope that 8 concurrent jobs will strick a good
balance between speed and resource consumption, considering that
test cases mostly wait.

That is still below what we do for Circle CI builds with 12 jobs
for distcheck tasks.