finish my first pass trough commits for the changelog
this might need more love in places, but I wanted to get my first
pass done and pushed before an ICE train arrives in Hamburg.

I suggest we aim for a complete and correct changelog first before we
rewrite/reformat it for the release- and upgrading notes.

Please review and feel free to fix/improve or add additional changes
which I left out or might have missed.