Use VCL caching also for backend requests and esi subrequests
The vcl member of (struct worker_priv) acts as a cache for a VCL reference:
In Req_Cleanup(), we call VCL_Recache() to (potentially) save a VCL
reference to (struct worker_priv) and we already use it in CNT_Embark().

We now extend this mechanism to backend requests and esi subrequests: On
the way in, we take the cached vcl reference if it is the right one (which
is highly likely except for vcl switching).

Likewise, on the way out of the ESI subrequest or backend request, we
now (potentially) save our VCL reference to the worker for reuse.

In total, this caching pontentially avoids taking vcl_mtx twice per
backend request and ESI subrequest.