Use CircleCI 'machine' executors with QEMU emulation to build packages
This way CircleCI could be used to build, test and create multiarch packages.

Alpine Linux aarch64 packaging is currently commented out because it fails while installing the dependencies. Needs more debugging, but the issue must be in the base Docker image because the same
Shell script works just fine on amd64