add back a local variable which makes flexelint grok the code
I did not understand when I committed
119b41747f9dd3360d1fe1ea1cb181c9aefb2e81 that the local variable was
required for flexelint to unterstand that this is not an out-of-bounds
access.

This quote of phk from the top level rant sais it all:

 * Do I need to tell you that static code analysis tools have a
 * really hard time coping with this, and that they give a lot of
 * false negatives which confuse people ?

So true. Sorry for going a smaller but almost full circle here, I
started with good intentions and now all that's left is the desire to
leave the code at least a little cleaner as I found it.