Centralize control over Range headers in the range VDP
This is an alternative approach to the ideas recorded in #3251: By
centralizing control over the Range-Request related headers, range
handling can be disabled by removing the range vdp from the response
filters.

By default, we now always push the range VDP as long as
http_range_support is on. If that is the case, behaviour is just the
same as before, but the Accept-Ranges response header is now set in the
range VDP's init callback.

The difference is that, if the range vdp is removed from the filters, no
Accept-Ranges header will be sent.

Note that if the range will not be active during delivery if it only
sets Accept-Ranges (by returning 1 from the init callback).

Fixes #3251