wssan: Track workspace reservations
This is a tiny bit more complicated because we have three WS_Reserve*()
functions with varying calling conventions, and as a bonus two more
WS_Release*() functions.

Just like WS_Alloc() and WS_Copy(), WS_ReserveAll() and WS_ReserveSize()
actually benefit from logic de-duplication between the two.

And there is one major change with WS_ReserveSize() in the context of
the workspace sanitizer: instead of reserving a pointer-aligned buffer
it now gives either zero or the exact number of requested bytes to the
caller. As a result ws->r no longer has alignment requirements and it's
ws_Release() that is in charge of guaranteeing that once updated, ws->f
maintains its alignment requirement.

This change is visible in r03131.vtc that now exhibits this match.

Because we can't predict the alignment requirements at reservation time,
we can use the align field of struct wa_alloc to reserve up to the size
of a pointer, which is an impossible outcome for plain allocations. We
can check at release time that we are indeed adjusting the tracking of a
reservation.