wssan: Unwind allocations
The ws_alloc leak that was intentionally left alone so far is finally
plugged.

It is no longer possible to assume that an allocation is a safe reset
spot: wssan helped spot that std.ip() was not respecting the snapshot
and reset contract. So in addition to the original goal of detecting
buffer overflows, it turns out wssan can do more.

A rollback of the workspace will opportunistically reallocate a new
workspace sanitizer.

Speaking of rollbacks, we need to perform them comprehensively in order
to combine wssan with other sanitizers like lsan. Workspace allocations
happen in preallocated space and individual allocations aren't freed.
With wssan however, we allocate the sanitizer at the very beginning of
the workspace, but keep track of individual allocations on the heap. We
therefore need to unwind all allocations at the end of a task, which is
cheap without wssan. Since WS_Rollback() is the only interface that
allows us to bring a workspace "back to the start" but as a side effect
it clears the overflow marker, we need to make sure that the final
rollback happens once we are really done with the workspace.