ws: Move the workspace panic dump to cache_ws.c
The goal is to avoid direct field access inside struct ws outside of
cache_ws.c and open the possibility to perform a hexdump of a corrupted
allocation in the future, when wssan panics while checking red zones.

This effectively completes the ban of direct field access to the front,
start and reservation pointers outside of cache_ws.c, and since the
devil is in the details cache_http.c directly touches the id field and
vmod_vtc also accesses the 3 aforementioned pointers.