ws: Ban direct access to the workspace id
From this point on, only cache_ws.c fiddles with struct ws, which needs
to remain visible in order to be embeddable in other data structures. We
have an API covering all use cases in tree, except vmod_vtc operations
that violate the contract to provide a rudimentary dump in VCL for test
purposes.

Inline functions provided in headers aren't considered a problem here.