Move synth reason between workspaces from vcl_pipe
In vcl_pipe req is read-only and we operate bereq as a busy object, as a
result we use the busy object's workspace to back VCL expressions.  When
we use the synth transition we may build a reason field on the wrong
workspace since vcl_synth carries on in a client context.

It is possible to notice that the reason string belongs to the backend
workspace and make a copy inside the client workspace before the former
is released.

Fixes #3329